cleanup push update

+++ src/heredocs_utils.c
@@ -6,26 +6,30 @@
 /*   By: vagarcia <vagarcia@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 11:19:03 by vagarcia          #+#    #+#             */
-/*   Updated: 2025/04/14 11:21:33 by vagarcia         ###   ########.fr       */
+/*   Updated: 2025/04/27 11:05:00 by vagarcia         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/minishell.h"
 
-const char	*get_random_temp_name(void)
+char	*get_random_temp_name(void)
 {
-	static char		temp_name[32];
+	char			*temp_name;
 	int				fd;
 	unsigned char	random_bytes[6];
 	int				i;
 
-	ft_strlcpy(temp_name, HERE_TEMP, sizeof(temp_name));
+	temp_name = xmalloc(32);
+	if (!temp_name)
+		return (NULL);
+	ft_strlcpy(temp_name, HERE_TEMP, 32);
 	fd = open("/dev/urandom", O_RDONLY);
 	if (fd == -1 || read(fd, random_bytes, 6) != 6)
 	{
-		perror("get_random_temp_name");
+		perror("/dev/urandom");
 		if (fd != -1)
 			close(fd);
+		safe_free(temp_name);
 		return (NULL);
 	}
 	close(fd);
@@ -41,42 +45,31 @@ char	*clean_empty_expansions(char *arg)
 	char	*result;
 	int		i;
 	int		j;
-	
+
+	if (!arg)
+		return (NULL);
 	result = xmalloc(ft_strlen(arg) + 1);
 	if (!result)
-	return (arg);
+		return (arg);
 	i = 0;
 	j = 0;
 	while (arg[i])
 	{
-		if (arg[i] == '$' && (arg[i + 1] == ' ' || arg[i + 1] == '\0' || arg[i
-			+ 1] == '$' || arg[i + 1] == '/'))
+		if (arg[i] == '$' && (arg[i + 1] == ' ' || arg[i + 1] == '\0'
+				|| arg[i + 1] == '$' || arg[i + 1] == '/'))
 			i++;
-			else
+		else
 			result[j++] = arg[i++];
 	}
 	result[j] = '\0';
 	if (j < (int)ft_strlen(arg))
 	{
-		xfree(arg);
+		free(arg);
 		return (result);
 	}
-	xfree(result);
+	free(result);
 	return (arg);
 }
-
-// char	*process_line(char *arg, t_cmd *cmd, bool expand_vars)
-// {
-// 	char	*expanded_line;
-
-// 	if (!expand_vars)
-// 		return (arg);
-// 	expanded_line = process_argument(arg, cmd->shell);
-// 	if (!expanded_line)
-// 		return (arg);
-// 	arg = handle_expanded_line(arg, expanded_line);
-// 	return (clean_empty_expansions(arg));
-// }
 
 char	*remove_quotes(char *lim)
 {
@@ -84,6 +77,8 @@ char	*remove_quotes(char *lim)
 	int		i;
 	int		j;
 
+	if (!lim)
+		return (NULL);
 	new_lim = xmalloc(sizeof(char) * (ft_strlen(lim) + 1));
 	if (!new_lim)
 		return (NULL);
@@ -96,7 +91,7 @@ char	*remove_quotes(char *lim)
 		i++;
 	}
 	new_lim[j] = '\0';
-	xfree(lim);
+	free(lim);
 	return (new_lim);
 }
 
@@ -105,16 +100,21 @@ void	cleanup_heredocs(t_shell *shell)
 	t_redir	*redir;
 	t_redir	*next;
 
+	if (!shell || !shell->cmd)
+		return ;
 	redir = shell->cmd->redirs;
 	while (redir && redir->type == REDIR_HEREDOC)
 	{
+		next = redir->next;
 		if (redir->type == REDIR_HEREDOC && redir->file
 			&& unlink(redir->file) == -1)
 			perror("unlink");
-		xfree(redir->file);
-		next = redir->next;
-		xfree(redir);
+		if (redir->file)
+			safe_free(redir->file);
+		if (redir->prefile)
+			safe_free(redir->prefile);
+		free(redir);
 		redir = next;
 	}
-	shell->cmd->redirs = NULL;
+	shell->cmd->redirs = redir;
 }
